refactor(Example NextJS): Update blog post metadata handling and reorganize imports
