+++ examples/nextjs/src/hooks/isEditMode.js
