+++ examples/nextjs/src/pages/DetailPage.js
@@ -1,40 +1,55 @@
 "use client";
 
 import { useEffect, useState } from "react";
+import Image from "next/image";
 
 import { enableBlockEditorInline } from "@dotcms/uve";
 import { DotCMSBlockEditorRenderer, useEditableDotCMSPage } from "@dotcms/react/next";
 
-import { isEditMode } from "@/utils/isEditMode";
+import { isEditMode } from "@/hooks/isEditMode";
 import Footer from "@/components/layout/footer/footer";
 import Header from "@/components/layout/header/header";
 
 export function DetailPage({ pageContent }) {
-    const [twActives, setTwActives] = useState("");
+    const [twActives, setTwActives] = useState("prose lg:prose-xl prose-a:text-blue-600");
     const { pageAsset } = useEditableDotCMSPage(pageContent);
     const { urlContentMap } = pageAsset;
     const { blogContent } = urlContentMap || {};
 
     useEffect(() => {
         if (isEditMode()) {
-            setTwActives( "border-2 border-solid border-cyan-400 cursor-pointer");
+            setTwActives((prev) => `${prev} border-2 border-solid border-cyan-400 cursor-pointer`);
         }
     }, []);
 
     const handleClick = () => {
-        enableBlockEditorInline(urlContentMap, "blogContent");
+        if (isEditMode()) {
+            enableBlockEditorInline(urlContentMap, "blogContent");
+        }
     };
 
     return (
         <div className="flex flex-col gap-6 min-h-screen bg-slate-50">
             {pageAsset?.layout.header && <Header />}
+            <main className="flex flex-col gap-8 m-auto">
+                {urlContentMap?.image && (
+                    <div className="relative w-full h-80 overflow-hidden">
+                        <Image
+                            className="object-cover"
+                            src={urlContentMap.inode}
+                            fill={true}
+                            alt="Activity Image"
+                        />
+                    </div>
+                )}
 
-            <main className="flex flex-col gap-8 m-auto" onClick={handleClick}>
-                <DotCMSBlockEditorRenderer
-                    blocks={JSON.parse(blogContent)}
-                    className={twActives}
-                    customRenderers={customeRenderers}
-                />
+                <div onClick={handleClick}>
+                    <DotCMSBlockEditorRenderer
+                        blocks={JSON.parse(blogContent)}
+                        className={twActives}
+                        customRenderers={customeRenderers}
+                    />
+                </div>
             </main>
 
             {/* {pageAsset?.layout.footer && <Footer />} */}
